+++ src/main/kotlin/com/bibireden/data_attributes/config/models/OverridesConfigModel.kt
@@ -2,6 +2,7 @@ package com.bibireden.data_attributes.config.models
 
 import com.bibireden.data_attributes.DataAttributes
 import com.bibireden.data_attributes.api.attribute.StackingFormula
+import com.bibireden.data_attributes.api.attribute.AttributeFormat
 import com.bibireden.data_attributes.mutable.MutableEntityAttribute
 import io.wispforest.endec.Endec
 import io.wispforest.endec.impl.StructEndecBuilder
@@ -36,7 +37,9 @@ class OverridesConfigModel {
         @JvmField
         var max_fallback: Double = 20.0,
         @JvmField
-        var formula: StackingFormula = StackingFormula.Flat
+        var formula: StackingFormula = StackingFormula.Flat,
+        @JvmField
+        var format: AttributeFormat = AttributeFormat.Whole,
     ) {
         /** Calls an override of an `MutableEntityAttribute`. */
         fun override(attribute: EntityAttribute) {
@@ -52,7 +55,8 @@ class OverridesConfigModel {
                 Endec.DOUBLE.optionalFieldOf("smoothness", { it.smoothness }, 0.01),
                 Endec.DOUBLE.optionalFieldOf("min_fallback", { it.min_fallback }, 0.0),
                 Endec.DOUBLE.optionalFieldOf("max_fallback", { it.max_fallback }, 1_000_000.0),
-                StackingFormula.ENDEC.fieldOf("formula") { it.formula },
+                StackingFormula.ENDEC.optionalFieldOf("formula", { it.formula }, StackingFormula.Flat),
+                AttributeFormat.ENDEC.optionalFieldOf("format", { it.format }, AttributeFormat.Whole),
                 ::AttributeOverride,
             )
         }